// webapp/resources/sap/esh/search/ui/sinaNexTS/sina/ConditionType.d.ts
declare module "sap/esh/search/ui/sinaNexTS/sina/ConditionType" {
    /*!
     * SAPUI5
     * Copyright (c) 2025 SAP SE or an SAP affiliate company. All rights reserved.
     *
     */
    enum ConditionType {
        Simple = "Simple",
        Complex = "Complex"
    }
}
//# sourceMappingURL=ConditionType.d.ts.map